test(validation): move the write destination tests to their own file
Adding the link-to-root cases pushed the predecessor test file to 533 lines,
past the 500 line ceiling. The split is along the seam that was already there:
one file attacks the numbers the guard compares, the other attacks where the
write lands. Fifty tests before, fifty after.

Refs #4233
